+++ src/Phone.cpp
@@ -8,12 +8,12 @@
 #include "Phone.h"
 
 using namespace std;
-
-	double detectFrequency(vector<double> pcm,double sampleRate){
+/*
+	float detectFrequency(vector<float> pcm,float sampleRate){
 		int length = pcm.size();
 		int minPeriod = floor(sampleRate/1046.5);//maximum frequency humans can sing
 		int maxPeriod = floor(sampleRate/82.407);//minimum frequency humans can sing
-		vector<double> errors(maxPeriod-minPeriod,0);
+		vector<float> errors(maxPeriod-minPeriod,0);
 		for(int period=minPeriod; period<maxPeriod; period++){
 			int periods = length/period;//number of periods that fit in the sample
 			for(int periodsIn=0; periodsIn<periods-1; periodsIn++){
@@ -23,31 +23,32 @@ using namespace std;
 			}
 			errors[period-minPeriod] /= (periods-1)*period;
 		}
-		return double(sampleRate)/(minPeriod+distance( errors.begin(), max_element(errors.begin(),errors.end()) ));
+		return float(sampleRate)/(minPeriod+distance( errors.begin(), max_element(errors.begin(),errors.end()) ));
 	}
-	double detectEnergy(vector<double> pcm){
-		double energy =0.;
+	float detectEnergy(vector<float> pcm){
+		float energy =0.;
 		for(int i=0; i<pcm.size(); i++){
 			energy += pcm[i]*pcm[i];
 		}
 		return energy;
 	}
+	*/
 Phone::Phone(
-		vector<double> pcm,
-		double consonantTime, double preutterTime, double overlapTime,
+		vector<float> pcm,
+		float consonantTime, float preutterTime, float overlapTime,
 		int windowOverlap, int windowSize, int sampleRate){
 
 	//initialize class variables
 	consonant = consonantTime;
 	preutter = preutterTime;
 	overlap =  overlapTime;
-	vector<double> vowelPart = vector<double>(
+	vector<float> vowelPart = vector<float>(
 		pcm.begin()+consonant*sampleRate,
 		pcm.end()
 	);
-	//double frequency = detectFrequency(vowelPart, sampleRate);
+	//float frequency = detectFrequency(vowelPart, sampleRate);
 	 /*
-	double powerroot = sqrt(detectEnergy(vowelPart))/pcm.size();
+	float powerroot = sqrt(detectEnergy(vowelPart))/pcm.size();
 	cerr<<powerroot<<endl;
 	for(int i=0; i<pcm.size(); i++){
 		pcm[i] /= powerroot;
@@ -60,14 +61,14 @@ Phone::Phone(
 	}
 	sample = Speech(Sound(pcm, windowOverlap, windowSize, sampleRate));
 }
-Phone::Phone(Speech speechSample, double consonantTime, double preutterTime, double overlapTime){
+Phone::Phone(Speech speechSample, float consonantTime, float preutterTime, float overlapTime){
 	consonant = consonantTime;
 	preutter = preutterTime;
 	overlap = overlapTime;
 	if(overlap<0){
 		overlap *= -1;
 		sample = Speech(
-			speechSample.startToSound(0).compatibleSound(vector<double>(overlap*speechSample.sampleRate))
+			speechSample.startToSound(0).compatibleSound(vector<float>(overlap*speechSample.sampleRate))
 		);
 		sample.add(speechSample,0);
 		preutter += overlap;
@@ -78,21 +79,21 @@ Phone::Phone(Speech speechSample, double consonantTime, double preutterTime, dou
 Phone::Phone(int overlapFactor, int windowLength, int sampleRate){
 	consonant = preutter = overlap = 0;
 	sample = Speech(Sound(
-		vector<double>(windowLength),
+		vector<float>(windowLength),
 		overlapFactor, windowLength, sampleRate
 	));
 }
 
 
-double Phone::getConsonant(){ return consonant; }
-double Phone::getPreutter (){ return preutter; }
-double Phone::getOverlap  (){ return overlap; }
+float Phone::getConsonant(){ return consonant; }
+float Phone::getPreutter (){ return preutter; }
+float Phone::getOverlap  (){ return overlap; }
 
 
 Phone::Phone(istream& filestream){
-	this->consonant = fileio::read(filestream,double(0));
-	this->preutter  = fileio::read(filestream,double(0));
-	this->overlap   = fileio::read(filestream,double(0));
+	this->consonant = fileio::read(filestream,float(0));
+	this->preutter  = fileio::read(filestream,float(0));
+	this->overlap   = fileio::read(filestream,float(0));
 	this->sample = Speech(filestream);
 }
 void Phone::write(ostream& filestream){
