+++ src/Speech.h
@@ -12,30 +12,30 @@ class Speech{
 		int hops;
 		
 		//time-based data
-		std::vector< std::vector<double> > magnitudes;
-		std::vector< std::vector<double> > freqDisplacements;
-		std::vector<double> frequencies;
+		std::vector< std::vector<float> > magnitudes;
+		std::vector< std::vector<float> > freqDisplacements;
+		std::vector<float> frequencies;
 
 		//helper methods/data
-		std::vector<double> remainder;
+		std::vector<float> remainder;
 		Sound toSound(int endHop);
 	public:
 		//public knowledge
-		double sampleRate;
-		double duration;
+		float sampleRate;
+		float duration;
 
 		//creation, addition, & destruction
 		Speech(Sound sample=Sound());
-		void add(Speech addee, double overlap);
-		Sound startToSound(double endTime);
-		void crop(double starttime, double endtime);
-		std::vector<double> pop(double popLength);
-		std::vector<double> synthesize();
+		void add(Speech addee, float overlap);
+		Sound startToSound(float endTime);
+		void crop(float starttime, float endtime);
+		std::vector<float> pop(float popLength);
+		std::vector<float> synthesize();
 
 		//modifying
-		void stretch(double start, double end, double nuvolength);
-		void amplify(std::function<double (double)> freqs);
-		void transpose(std::function<double (double)> frequency, double endTime);
+		void stretch(float start, float end, float nuvolength);
+		void amplify(std::function<float (float)> freqs);
+		void transpose(std::function<float (float)> frequency, float endTime);
 
 		//for reading & writing to disc
 		Speech(std::istream& filestream);
